Added a new section and fixed various bugs

+++ client/src/components/Groups.js
@@ -19,7 +19,14 @@ import Typography from '@mui/material/Typography';
 import PhoneInTalkIcon from '@mui/icons-material/PhoneInTalk';
 import CancelIcon from '@mui/icons-material/Cancel';
 
-import { bluegrey, richBlack, light, medium, dark, deepDark } from './colors';
+import {
+    bluegrey,
+    richBlack,
+    light,
+    medium,
+    dark,
+    deepDark,
+} from '../utils/colors';
 import { useHMSActions } from '@100mslive/hms-video-react';
 import {
     startLoadingAction,
@@ -27,7 +34,7 @@ import {
     notifyAction,
 } from '../actions/actions';
 
-function Groups({ themeChange, mode }) {
+function Groups({ mode }) {
     const navigate = useNavigate();
     const dispatch = useDispatch();
     const hmsActions = useHMSActions();
@@ -174,6 +181,9 @@ function Groups({ themeChange, mode }) {
         );
         if (response.data.success) {
             setModalVisible(false);
+            setTitle('');
+            setDescription('');
+            setCoverImgURL(null);
             window.open(`${process.env.REACT_APP_BASE_URL}/groups`, '_self');
         } else {
             alert('Something went wrong, please try again');
@@ -181,23 +191,31 @@ function Groups({ themeChange, mode }) {
     };
 
     const deleteSpace = async (roomId) => {
-        const result = await axios.delete(
-            `${process.env.REACT_APP_SERVER_URL}/api/rooms/delete/${roomId}`
+        const choice = window.confirm(
+            'Are you sure you want to delete this space?'
         );
-        if (result.data.success) {
-            window.location.reload();
-            alert('Space deleted successfully');
-        } else {
-            alert('Something went wrong, please try again');
+        if (choice) {
+            const result = await axios.delete(
+                `${process.env.REACT_APP_SERVER_URL}/api/rooms/delete/${roomId}`
+            );
+            if (result.data.success) {
+                window.location.reload();
+            } else {
+                alert('Something went wrong, please try again');
+            }
         }
     };
 
     return (
         <Box
             sx={{
-                minHeight: '100vh',
+                overflowY: 'auto',
+                mt: '75px',
+                height: 'calc(100vh - 75px)',
+                maxHeight: 'calc(100vh - 75px)',
                 backgroundColor: mode === 'light' ? light : bluegrey,
                 padding: '5rem',
+                pt: 0,
             }}
         >
             <Typography
@@ -206,7 +224,6 @@ function Groups({ themeChange, mode }) {
                 sx={{
                     color: mode === 'light' ? deepDark : light,
                     margin: '2rem',
-                    fontFamily: 'Poppins, Work Sans',
                     fontWeight: 'bold',
                     fontSize: '3rem',
                     textAlign: 'center',
@@ -245,8 +262,10 @@ function Groups({ themeChange, mode }) {
             <Box
                 sx={{
                     display: 'grid',
-                    gridTemplateColumns: 'repeat(3, 1fr)',
-                    gridGap: '1rem',
+                    borderRadius: '10px',
+                    gridTemplateColumns: 'repeat(3, minmax(280px, 1fr))',
+                    gap: '24px 24px',
+                    gridAutoFlow: 'dense',
                 }}
             >
                 {spaces &&
@@ -388,8 +407,7 @@ function Groups({ themeChange, mode }) {
                         left: '50%',
                         transform: 'translate(-50%, -50%)',
                         minWidth: 600,
-                        maxWidth: 700,
-                        height: 670,
+                        maxHeight: '700px',
                         backgroundColor: mode === 'light' ? light : bluegrey,
                         boxShadow: 24,
                         borderRadius: '10px',
@@ -409,7 +427,11 @@ function Groups({ themeChange, mode }) {
                         }}
                         cursor='pointer'
                         onClick={() => {
+                            setTitle('');
+                            setDescription('');
+                            setCoverImgURL('');
                             setModalVisible(false);
+                            return;
                         }}
                     />
                     <Typography
@@ -469,36 +491,25 @@ function Groups({ themeChange, mode }) {
                             }}
                         />
                         {coverImgURL && (
-                            <>
-                                <img
-                                    style={{
-                                        objectFit: 'fill',
-                                        height: '350px',
-                                        width: '455px',
-                                        alignSelf: 'center',
-                                        position: 'relative',
-                                    }}
-                                    alt='loading ...'
-                                    src={coverImgURL}
-                                />
-                                <CancelIcon
-                                    sx={{
-                                        position: 'absolute',
-                                        top: 210,
-                                        right: 80,
-                                    }}
-                                    cursor='pointer'
-                                    onClick={() => {
-                                        setCoverImgURL(null);
-                                    }}
-                                />
-                            </>
+                            <img
+                                style={{
+                                    objectFit: 'fill',
+                                    maxHeight: '300px',
+                                    width: '455px',
+                                    alignSelf: 'center',
+                                    position: 'relative',
+                                }}
+                                alt='loading ...'
+                                src={coverImgURL}
+                            />
                         )}
                         <Button
+                            disableElevation
                             color='success'
                             variant='contained'
                             sx={{
                                 mt: 1,
+                                alignSelf: 'center',
                                 backgroundColor:
                                     mode === 'light' ? medium : light,
                                 color: 'black',
@@ -514,13 +525,11 @@ function Groups({ themeChange, mode }) {
                         <Button
                             color='success'
                             variant='contained'
+                            disableElevation
                             sx={{
-                                position: 'absolute',
-                                bottom: 10,
-                                right: 15,
+                                mt: 1,
                                 mb: 1,
                                 alignSelf: 'flex-end',
-                                // mr: 7,
                                 backgroundColor:
                                     mode === 'light' ? medium : light,
                                 color: 'black',
